Add RollingTrend primitives and refactor some util methods (#2170)
* add RollingTrend primitives and refactor some util methods

* release notes

* get RollingTrend to work, separate util methods into own file

* example doc string

* unit tests

* fix unit test

* pr comments

* remove rollingtrend from non uniform test

* unit tests + extra fixes for pre commit lint

* I don't like doctests

* fix code cov

* run code cov in 3.8

* add time utils test and api reference

* spelling

* cap woodwork requirement for now

* uncap woodwork

* cover calculate_trend

* add timestamp case

add timestamp case

* timedelta no time change
