+++ .github/workflows/unit_tests_with_latest_deps.yaml
@@ -50,13 +50,13 @@ jobs:
         run: |
           cd unpacked_sdist
           pytest featuretools/ -n 2
-      - if: ${{ matrix.libraries == 'spark' }}
+      - if: ${{ matrix.python_version == 3.8 && matrix.libraries == 'spark' }}
         name: Run unit tests with code coverage
         run: |
           cd unpacked_sdist/
           coverage erase
           pytest featuretools/ -n 2 --cov=featuretools --cov-config=../pyproject.toml --cov-report=xml:../coverage.xml
-      - if: ${{ matrix.libraries == 'spark' }}
+      - if: ${{ matrix.python_version == 3.8 && matrix.libraries == 'spark' }}
         name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
         with:
