+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-exclude: ^LICENSES/|\.(html|csv|svg|md)$
+exclude: ^LICENSE/|\.(html|csv|svg|md)$
 default_stages: [commit]
 repos:
   - repo: https://github.com/kynan/nbstripout
@@ -16,7 +16,7 @@ repos:
   - repo: https://github.com/MarcoGorelli/absolufy-imports
     rev: v0.3.1
     hooks:
-    -   id: absolufy-imports
+      - id: absolufy-imports
         files: ^featuretools/
   - repo: https://github.com/asottile/add-trailing-comma
     rev: v2.2.3
