+++ docs/source/api_reference.rst
@@ -157,6 +157,7 @@ Rolling Transform Primitives
     RollingMean
     RollingMin
     RollingSTD
+    RollingTrend
 
 NaturalLanguage Transform Primitives
 ************************************
