+++ featuretools/primitives/standard/aggregation_primitives.py
@@ -1,4 +1,4 @@
-from datetime import datetime, timedelta
+from datetime import datetime
 
 import numpy as np
 import pandas as pd
@@ -14,7 +14,7 @@ from woodwork.logical_types import (
 )
 
 from featuretools.primitives.base.aggregation_primitive_base import AggregationPrimitive
-from featuretools.utils import convert_time_units
+from featuretools.utils import calculate_trend, convert_time_units
 from featuretools.utils.gen_utils import Library
 
 
@@ -772,60 +772,10 @@ class Trend(AggregationPrimitive):
 
     def get_function(self, agg_type=Library.PANDAS):
         def pd_trend(y, x):
-            df = pd.DataFrame({"x": x, "y": y}).dropna()
-            if df.shape[0] <= 2:
-                return np.nan
-            if isinstance(df["x"].iloc[0], (datetime, pd.Timestamp)):
-                x = convert_datetime_to_floats(df["x"])
-            else:
-                x = df["x"].values
 
-            if isinstance(df["y"].iloc[0], (datetime, pd.Timestamp)):
-                y = convert_datetime_to_floats(df["y"])
-            elif isinstance(df["y"].iloc[0], (timedelta, pd.Timedelta)):
-                y = convert_timedelta_to_floats(df["y"])
-            else:
-                y = df["y"].values
-
-            x = x - x.mean()
-            y = y - y.mean()
-
-            # prevent divide by zero error
-            if len(np.unique(x)) == 1:
-                return 0
-
-            # consider scipy.stats.linregress for large n cases
-            coefficients = np.polyfit(x, y, 1)
-
-            return coefficients[0]
+            return calculate_trend(pd.Series(data=y.values, index=x.values))
 
         return pd_trend
-
-
-def convert_datetime_to_floats(x):
-    first = int(x.iloc[0].value * 1e-9)
-    x = pd.to_numeric(x).astype(np.float64).values
-    dividend = find_dividend_by_unit(first)
-    x *= 1e-9 / dividend
-    return x
-
-
-def convert_timedelta_to_floats(x):
-    first = int(x.iloc[0].total_seconds())
-    dividend = find_dividend_by_unit(first)
-    x = pd.TimedeltaIndex(x).total_seconds().astype(np.float64) / dividend
-    return x
-
-
-def find_dividend_by_unit(time):
-    """Finds whether time best corresponds to a value in
-    days, hours, minutes, or seconds.
-    """
-    for dividend in [86400, 3600, 60]:
-        div = time / dividend
-        if round(div) == div:
-            return dividend
-    return 1
 
 
 class Entropy(AggregationPrimitive):
