+++ featuretools/tests/primitive_tests/test_rolling_primitive.py
@@ -1,5 +1,6 @@
 from unittest.mock import patch
 
+import numpy as np
 import pandas as pd
 import pytest
 
@@ -9,8 +10,9 @@ from featuretools.primitives import (
     RollingMean,
     RollingMin,
     RollingSTD,
+    RollingTrend,
+    roll_series_with_gap,
 )
-from featuretools.primitives.utils import _roll_series_with_gap
 from featuretools.tests.primitive_tests.utils import get_number_from_offset
 
 
@@ -29,25 +31,20 @@ def test_rolling_max(min_periods, window_length, gap, rolling_series_pd):
     window_length_num = get_number_from_offset(window_length)
     # Since we're using a uniform series we can check correctness using numeric parameters
     expected_vals = (
-        _roll_series_with_gap(
-            rolling_series_pd,
-            window_length_num,
-            gap=gap_num,
-            min_periods=min_periods,
+        roll_series_with_gap(
+            rolling_series_pd, window_length_num, gap=gap_num, min_periods=min_periods
         )
         .max()
         .values
     )
 
     primitive_instance = RollingMax(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
 
     actual_vals = pd.Series(
-        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values)),
+        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
     )
 
     # Since min_periods of 0 is the same as min_periods of 1
@@ -73,25 +70,20 @@ def test_rolling_min(min_periods, window_length, gap, rolling_series_pd):
 
     # Since we're using a uniform series we can check correctness using numeric parameters
     expected_vals = (
-        _roll_series_with_gap(
-            rolling_series_pd,
-            window_length_num,
-            gap=gap_num,
-            min_periods=min_periods,
+        roll_series_with_gap(
+            rolling_series_pd, window_length_num, gap=gap_num, min_periods=min_periods
         )
         .min()
         .values
     )
 
     primitive_instance = RollingMin(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
 
     actual_vals = pd.Series(
-        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values)),
+        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
     )
 
     # Since min_periods of 0 is the same as min_periods of 1
@@ -117,25 +109,20 @@ def test_rolling_mean(min_periods, window_length, gap, rolling_series_pd):
 
     # Since we're using a uniform series we can check correctness using numeric parameters
     expected_vals = (
-        _roll_series_with_gap(
-            rolling_series_pd,
-            window_length_num,
-            gap=gap_num,
-            min_periods=min_periods,
+        roll_series_with_gap(
+            rolling_series_pd, window_length_num, gap=gap_num, min_periods=min_periods
         )
         .mean()
         .values
     )
 
     primitive_instance = RollingMean(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
 
     actual_vals = pd.Series(
-        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values)),
+        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
     )
 
     # Since min_periods of 0 is the same as min_periods of 1
@@ -161,25 +148,20 @@ def test_rolling_std(min_periods, window_length, gap, rolling_series_pd):
 
     # Since we're using a uniform series we can check correctness using numeric parameters
     expected_vals = (
-        _roll_series_with_gap(
-            rolling_series_pd,
-            window_length_num,
-            gap=gap_num,
-            min_periods=min_periods,
+        roll_series_with_gap(
+            rolling_series_pd, window_length_num, gap=gap_num, min_periods=min_periods
         )
         .std()
         .values
     )
 
     primitive_instance = RollingSTD(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
 
     actual_vals = pd.Series(
-        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values)),
+        primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
     )
 
     # Since min_periods of 0 is the same as min_periods of 1
@@ -208,7 +190,7 @@ def test_rolling_count(window_length, gap, rolling_series_pd):
     window_length_num = get_number_from_offset(window_length)
 
     expected_vals = (
-        _roll_series_with_gap(
+        roll_series_with_gap(
             rolling_series_pd,
             window_length_num,
             gap=gap_num,
@@ -219,9 +201,7 @@ def test_rolling_count(window_length, gap, rolling_series_pd):
     )
 
     primitive_instance = RollingCount(
-        window_length=window_length,
-        gap=gap,
-        min_periods=window_length_num,
+        window_length=window_length, gap=gap, min_periods=window_length_num
     )
     primitive_func = primitive_instance.get_function()
 
@@ -229,11 +209,10 @@ def test_rolling_count(window_length, gap, rolling_series_pd):
 
     num_nans = gap_num + window_length_num - 1
     assert actual_vals.isna().sum() == num_nans
-    # RollingCount will not match the exact _roll_series_with_gap call,
+    # RollingCount will not match the exact roll_series_with_gap call,
     # because it handles the min_periods difference within the primitive
     pd.testing.assert_series_equal(
-        pd.Series(expected_vals).iloc[num_nans:],
-        actual_vals.iloc[num_nans:],
+        pd.Series(expected_vals).iloc[num_nans:], actual_vals.iloc[num_nans:]
     )
 
 
@@ -243,16 +222,10 @@ def test_rolling_count(window_length, gap, rolling_series_pd):
 )
 @pytest.mark.parametrize("window_length, gap", [("5d", "2d"), (5, 2)])
 def test_rolling_count_primitive_min_periods_nans(
-    window_length,
-    gap,
-    min_periods,
-    expected_num_nams,
-    rolling_series_pd,
+    window_length, gap, min_periods, expected_num_nams, rolling_series_pd
 ):
     primitive_instance = RollingCount(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
     vals = pd.Series(primitive_func(rolling_series_pd.index))
@@ -266,16 +239,10 @@ def test_rolling_count_primitive_min_periods_nans(
 )
 @pytest.mark.parametrize("window_length, gap", [("5d", "0d"), (5, 0)])
 def test_rolling_count_with_no_gap(
-    window_length,
-    gap,
-    min_periods,
-    expected_num_nams,
-    rolling_series_pd,
+    window_length, gap, min_periods, expected_num_nams, rolling_series_pd
 ):
     primitive_instance = RollingCount(
-        window_length=window_length,
-        gap=gap,
-        min_periods=min_periods,
+        window_length=window_length, gap=gap, min_periods=min_periods
     )
     primitive_func = primitive_instance.get_function()
     vals = pd.Series(primitive_func(rolling_series_pd.index))
@@ -284,8 +251,72 @@ def test_rolling_count_with_no_gap(
 
 
 @pytest.mark.parametrize(
+    "window_length, gap, expected_vals",
+    [
+        (3, 0, [np.nan, np.nan, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]),
+        (
+            4,
+            1,
+            [np.nan, np.nan, np.nan, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+        ),
+        (
+            "5d",
+            "7d",
+            [
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                np.nan,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+            ],
+        ),
+        (
+            "5d",
+            "0d",
+            [np.nan, np.nan, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+        ),
+    ],
+)
+def test_rolling_trend(window_length, gap, expected_vals, rolling_series_pd):
+    primitive_instance = RollingTrend(window_length=window_length, gap=gap)
+
+    actual_vals = primitive_instance(rolling_series_pd.index, rolling_series_pd.values)
+
+    pd.testing.assert_series_equal(pd.Series(expected_vals), pd.Series(actual_vals))
+
+
+def test_rolling_trend_window_length_less_than_three(rolling_series_pd):
+    primitive_instance = RollingTrend(window_length=2)
+
+    vals = primitive_instance(rolling_series_pd.index, rolling_series_pd.values)
+
+    for v in vals:
+        assert np.isnan(v)
+
+
+@pytest.mark.parametrize(
     "primitive",
-    [RollingCount, RollingMax, RollingMin, RollingMean],
+    [
+        RollingCount,
+        RollingMax,
+        RollingMin,
+        RollingMean,
+    ],
 )
 def test_rolling_primitives_non_uniform(primitive):
     # When the data isn't uniform, this impacts the number of values in each rolling window
@@ -298,7 +329,7 @@ def test_rolling_primitives_non_uniform(primitive):
 
     # Should match RollingCount exactly and have same nan values as other primitives
     expected_series = pd.Series(
-        [None, 1, 2] + [None, 1, 1, 1] + [None, 1, 2, 3, 3, 3, 3],
+        [None, 1, 2] + [None, 1, 1, 1] + [None, 1, 2, 3, 3, 3, 3]
     )
 
     primitive_instance = primitive(window_length="3d", gap="1d")
@@ -307,7 +338,7 @@ def test_rolling_primitives_non_uniform(primitive):
         pd.testing.assert_series_equal(rolled_series, expected_series)
     else:
         rolled_series = pd.Series(
-            primitive_instance(no_freq_series.index, pd.Series(no_freq_series.values)),
+            primitive_instance(no_freq_series.index, pd.Series(no_freq_series.values))
         )
         pd.testing.assert_series_equal(expected_series.isna(), rolled_series.isna())
 
@@ -334,28 +365,51 @@ def test_rolling_std_non_uniform():
             1.0,
             1.0,
             1.0,
-        ],
+        ]
     )
 
     primitive_instance = RollingSTD(window_length="3d", gap="1d")
     rolled_series = pd.Series(
-        primitive_instance(no_freq_series.index, pd.Series(no_freq_series.values)),
+        primitive_instance(no_freq_series.index, pd.Series(no_freq_series.values))
     )
 
     pd.testing.assert_series_equal(rolled_series, expected_series)
 
 
+def test_rolling_trend_non_uniform():
+    datetimes = (
+        list(pd.date_range(start="2017-01-01", freq="1d", periods=3))
+        + list(pd.date_range(start="2017-01-10", freq="2d", periods=4))
+        + list(pd.date_range(start="2017-01-22", freq="1d", periods=7))
+    )
+    no_freq_series = pd.Series(range(len(datetimes)), index=datetimes)
+    expected_series = pd.Series(
+        [None, None, None]
+        + [None, None, None, None]
+        + [
+            None,
+            None,
+            None,
+            1.0,
+            1.0,
+            1.0,
+            1.0,
+        ]
+    )
+    primitive_instance = RollingTrend(window_length="3d", gap="1d")
+    rolled_series = pd.Series(
+        primitive_instance(no_freq_series.index, pd.Series(no_freq_series.values))
+    )
+    pd.testing.assert_series_equal(rolled_series, expected_series)
+
+
 @pytest.mark.parametrize(
     "primitive",
-    [RollingCount, RollingMax, RollingMin, RollingMean, RollingSTD],
+    [RollingCount, RollingMax, RollingMin, RollingMean, RollingSTD, RollingTrend],
 )
-@patch(
-    "featuretools.primitives.rolling_transform_primitive._apply_roll_with_offset_gap",
-)
+@patch("featuretools.primitives.rolling_transform_primitive.apply_roll_with_offset_gap")
 def test_no_call_to_apply_roll_with_offset_gap_with_numeric(
-    mock_apply_roll,
-    primitive,
-    rolling_series_pd,
+    mock_apply_roll, primitive, rolling_series_pd
 ):
     assert not mock_apply_roll.called
 
@@ -365,10 +419,7 @@ def test_no_call_to_apply_roll_with_offset_gap_with_numeric(
         pd.Series(primitive_func(rolling_series_pd.index))
     else:
         pd.Series(
-            primitive_func(
-                rolling_series_pd.index,
-                pd.Series(rolling_series_pd.values),
-            ),
+            primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
         )
 
     assert not mock_apply_roll.called
@@ -379,10 +430,7 @@ def test_no_call_to_apply_roll_with_offset_gap_with_numeric(
         pd.Series(primitive_func(rolling_series_pd.index))
     else:
         pd.Series(
-            primitive_func(
-                rolling_series_pd.index,
-                pd.Series(rolling_series_pd.values),
-            ),
+            primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
         )
 
     assert not mock_apply_roll.called
@@ -393,10 +441,7 @@ def test_no_call_to_apply_roll_with_offset_gap_with_numeric(
         pd.Series(primitive_func(rolling_series_pd.index))
     else:
         pd.Series(
-            primitive_func(
-                rolling_series_pd.index,
-                pd.Series(rolling_series_pd.values),
-            ),
+            primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
         )
 
     assert not mock_apply_roll.called
@@ -407,10 +452,7 @@ def test_no_call_to_apply_roll_with_offset_gap_with_numeric(
         pd.Series(primitive_func(rolling_series_pd.index))
     else:
         pd.Series(
-            primitive_func(
-                rolling_series_pd.index,
-                pd.Series(rolling_series_pd.values),
-            ),
+            primitive_func(rolling_series_pd.index, pd.Series(rolling_series_pd.values))
         )
 
     assert mock_apply_roll.called
