+++ featuretools/utils/api.py
@@ -7,5 +7,9 @@ from featuretools.utils.cli_utils import (
 )
 from featuretools.utils.entry_point import entry_point
 from featuretools.utils.gen_utils import make_tqdm_iterator
-from featuretools.utils.time_utils import convert_time_units, make_temporal_cutoffs
+from featuretools.utils.time_utils import (
+    calculate_trend,
+    convert_time_units,
+    make_temporal_cutoffs,
+)
 from featuretools.utils.trie import Trie
