+++ featuretools/utils/time_utils.py
@@ -1,3 +1,6 @@
+from datetime import datetime, timedelta
+
+import numpy as np
 import pandas as pd
 
 
@@ -77,3 +80,57 @@ def convert_time_units(secs, unit):
         raise ValueError("Invalid unit given, make sure it is plural")
 
     return secs / (unit_divs[unit])
+
+
+def convert_datetime_to_floats(x):
+    first = int(x.iloc[0].value * 1e-9)
+    x = pd.to_numeric(x).astype(np.float64).values
+    dividend = find_dividend_by_unit(first)
+    x *= 1e-9 / dividend
+    return x
+
+
+def convert_timedelta_to_floats(x):
+    first = int(x.iloc[0].total_seconds())
+    dividend = find_dividend_by_unit(first)
+    x = pd.TimedeltaIndex(x).total_seconds().astype(np.float64) / dividend
+    return x
+
+
+def find_dividend_by_unit(time):
+    """Finds whether time best corresponds to a value in
+    days, hours, minutes, or seconds.
+    """
+    for dividend in [86400, 3600, 60]:
+        div = time / dividend
+        if round(div) == div:
+            return dividend
+    return 1
+
+
+def calculate_trend(series):
+    df = pd.DataFrame({"x": series.index, "y": series.values}).dropna()
+    if df.shape[0] <= 2:
+        return np.nan
+    if isinstance(df["x"].iloc[0], (datetime, pd.Timestamp)):
+        x = convert_datetime_to_floats(df["x"])
+    else:
+        x = df["x"].values
+
+    if isinstance(df["y"].iloc[0], (datetime, pd.Timestamp)):
+        y = convert_datetime_to_floats(df["y"])
+    elif isinstance(df["y"].iloc[0], (timedelta, pd.Timedelta)):
+        y = convert_timedelta_to_floats(df["y"])
+    else:
+        y = df["y"].values
+
+    x = x - x.mean()
+    y = y - y.mean()
+
+    # prevent divide by zero error
+    if len(np.unique(x)) == 1:
+        return 0
+
+    # consider scipy.stats.linregress for large n cases
+    coefficients = np.polyfit(x, y, 1)
+    return coefficients[0]
