+++ pyproject.toml
@@ -23,7 +23,7 @@ skip = "__init__.py"
 multi_line_output = 3
 
 [tool.black]
-target-version = ['py10']
+target-version = ['py310']
 
 [tool.coverage.run]
 source = ["featuretools"]
